refactor: usePositionInfo hook
