+++ apps/Cybar/components/App/TradePanel/MinPositionInfo.tsx
@@ -1,7 +1,3 @@
-import BigNumber from "bignumber.js";
-import { useMemo } from "react";
-
-import { DEFAULT_PRECISION } from "@symmio/frontend-sdk/constants/misc";
 import { useCollateralToken } from "@symmio/frontend-sdk/constants/tokens";
 import { InputField } from "@symmio/frontend-sdk/types/trade";
 import {
@@ -13,11 +9,9 @@ import { useGetTokenWithFallbackChainId } from "@symmio/frontend-sdk/utils/token
 
 import useActiveWagmi from "@symmio/frontend-sdk/lib/hooks/useActiveWagmi";
 import {
-  useActiveMarket,
-  useActiveMarketPrice,
+  usePositionInfo,
   useSetTypedValue,
 } from "@symmio/frontend-sdk/state/trade/hooks";
-import { useLeverage } from "@symmio/frontend-sdk/state/user/hooks";
 
 import InfoItem from "components/InfoItem";
 
@@ -29,55 +23,12 @@ export default function MinPositionInfo() {
     COLLATERAL_TOKEN,
     chainId,
   );
-
-  const leverage = useLeverage();
-  const market = useActiveMarket();
-  const marketPrice = useActiveMarketPrice();
-  const [
+  const {
+    minPositionValue,
+    pricePrecision,
+    minPositionQuantity,
     outputTicker,
-    pricePrecision,
-    quantityPrecision,
-    minAcceptableQuoteValue,
-  ] = useMemo(
-    () =>
-      market
-        ? [
-            market.symbol,
-            market.pricePrecision,
-            market.quantityPrecision,
-            market.minAcceptableQuoteValue,
-            market.maxLeverage,
-          ]
-        : ["", DEFAULT_PRECISION, DEFAULT_PRECISION, 10],
-    [market],
-  );
-  const [minPositionValue, minPositionQuantity] = useMemo(() => {
-    // find maximum quantity between min quote value & minimum value base on quantity precision
-    const quantity = BigNumber.max(
-      toBN(minAcceptableQuoteValue)
-        .div(marketPrice)
-        .times(leverage)
-        .toFixed(quantityPrecision, RoundMode.ROUND_UP),
-      toBN(10)
-        .pow(quantityPrecision * -1)
-        .toFixed(quantityPrecision, RoundMode.ROUND_UP),
-    );
-    const value = toBN(quantity).times(marketPrice).div(leverage);
-
-    if (value.isNaN()) {
-      return ["-", "-"];
-    }
-    return [
-      value.toFixed(pricePrecision, RoundMode.ROUND_UP),
-      quantity.toFixed(quantityPrecision, RoundMode.ROUND_UP),
-    ];
-  }, [
-    leverage,
-    marketPrice,
-    minAcceptableQuoteValue,
-    pricePrecision,
-    quantityPrecision,
-  ]);
+  } = usePositionInfo();
 
   return (
     <InfoItem
